API fully functional. Project Version 2.0
Validate procedure now functions correctly, error was due to duplicate records that we're inserted in RDBMS thus causing api to error.

By deleting records from passwords, and sessions tables this indicates that the operation can only be done on one record and on muliple because the query returns mulitple records as there we're muliple with the same value it caused the api to error.

Noe ammended